requirements.txt
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
asm2vec==1.0.0
async-generator==1.10
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
awsebcli==3.20.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
botocore==1.20.98
browsermob-proxy==0.8.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
channels==3.0.3
chardet==4.0.0
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.1
django-allauth==0.44.0
django-colorfield==0.3.2
django-cors-headers==3.6.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
docutils==0.15.2
docx==0.2.4
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
future==0.16.0
hyperlink==21.0.0
identify==2.2.4
idna==2.10
incremental==17.5.0
inflection==0.5.1
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.2
jupyterlab-widgets==1.0.0
kiwisolver==1.2.0
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.2
matplotlib-inline==0.1.2
mistune==0.8.4
mpmath==1.2.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nodeenv==1.6.0
nose==1.3.7
notebook==6.4.0
numpy==1.19.2
oauthlib==3.1.0
olefile==0.47.dev4
opencv-python==4.5.3.56
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.5.9
pdf2image==1.14.0
pdfminer==20191125
pdfminer.six==20201018
pickleshare==0.7.5
Pillow==7.2.0
pre-commit==2.12.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.9.0
PyHamcrest==2.0.2
pyhwp==0.1b15
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.0
python-docx==0.8.11
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.10.0
readme2tex==0.0.1b2
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.3.0
soupsieve==2.1
sqlparse==0.4.1
sympy==1.8
termcolor==1.1.0
terminado==0.9.5
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.50.0
traitlets==5.0.5
Twisted==20.3.0
txaio==20.12.1
unoconv==0.9.0
uritemplate==4.1.1
urllib3==1.25.11
uuid==1.30
uuid1==1.0.1
virtualenv==20.4.6
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==1.0.1
wget==3.2
widgetsnbextension==3.5.1
zope.interface==5.2.0
